+++ rfsocinterface/analysis/__init__.py
@@ -1,1 +1,15 @@
 """Subpackage for all data analysis tools."""
+
+from rfsocinterface.analysis.psd import (
+    ComputeNoisePSD,
+    PlotPSD,
+    plot_psd_df_over_f,
+    plot_psd_dbc_hz,
+)
+
+__all__ = [
+    'ComputeNoisePSD',
+    'PlotPSD',
+    'plot_psd_dbc_hz',
+    'plot_psd_df_over_f',
+]
