+++ rfsocinterface/analysis/psd.py
@@ -1,5 +1,6 @@
 """Code for computing the noise PSD."""
 import pdb
+import logging
 
 from pathlib import Path
 from typing import Literal
@@ -10,430 +11,541 @@ from matplotlib.figure import Figure
 from scipy import signal
 from matplotlib.backends.backend_pdf import PdfPages
 from argparse import ArgumentParser
-import tables
 
 from kidpy3 import RawDataFile
 
-
 from rfsocinterface.core.data import (
-    flag_outliers,
     DataRoutine,
-    ProcessingStage,
+    PsdBasis,
+    ProcessedData,
+    decode_tone_indices,
+    flag_outliers,
+    register_routine,
 )
-from rfsocinterface.core.data.utils import PsdBasis
-from rfsocinterface.core.data.storage import ProcessedData
 from rfsocinterface.core.utils import DEFAULT_DATA_DIRECTORY, ensure_path, get_tod_template, ordinal, PERMISSIONS_ALL_FULL 
+
+
+_logger = logging.getLogger(__name__)
 
 XLIM = (0.1, 250)
 YLIM = (-110, -60)
 VALID_BASES = ['gp', 'iq', 'fd']
 
 
-def compute_noise_psd(
-    input_time_ordered_data: npt.NDArray,
-    timestamp: npt.NDArray,
-    chanmask: npt.NDArray | None=None,
-    nominal_block_length: float=1e100,
-    cut_time: float=0.0,
-) -> tuple[npt.NDArray, npt.NDArray, npt.NDArray, npt.NDArray]:
-    """Compute noise PSD.
+@register_routine
+class ComputeNoisePSD(DataRoutine):
+    name = 'ComputeNoisePSD'
+    version = '1.0.0'
 
-    input_time_ordered_data: 2 x N_res x N_sample or N_res x N_sample
-    timestamp: N_sample
-    chanmask: N_res
-    nominal_block_length: seconds
-    cut_time: seconds to cut from ends of data
-    """
-    first_dimension = input_time_ordered_data.shape[0] if input_time_ordered_data.ndim == 3 else 1
-    if first_dimension == 1:
-        input_data = input_time_ordered_data.reshape((1, *input_time_ordered_data.shape))
-    else:
-        input_data = input_time_ordered_data
-    if chanmask is None:
-        chanmask = np.ones_like(input_time_ordered_data[0, :, 0], dtype=int)
-        # chanmask[1000:] = 0  # This is a fix since these channels seem to be bad
+    def __init__(
+            self,
+            *bases: PsdBasis,
+            nominal_block_length: float=10,
+            cut_time: float=0.0,
+            selection_indices: npt.NDArray | str='all',
+    ):
+        super().__init__(
+            bases=bases,
+            nominal_block_length=nominal_block_length,
+            cut_time=cut_time,
+            selection_indices=selection_indices,
+        )
 
-    timestamp -= timestamp[0]
-    fs = 1. / np.median(np.diff(timestamp))
+    def inputs(self, pdata: ProcessedData) -> list[str]:
+        dsets = []
+        bases = self.params['bases']
+        for basis in bases:
+            match basis:
+                case PsdBasis.IQ:
+                    dsets.append('/vdsets/data_IQ')
+                case PsdBasis.GAIN_PHASE:
+                    dsets.append('/vdsets/data_gain_phase')
+                    dsets.append('/vdsets/carrier_amplitudes')
+                case PsdBasis.FREQ_DISS:
+                    dsets.append('/vdsets/data_freq_diss')
+                    dsets.append('/vdsets/tones')
+                case _:
+                    raise ValueError(f'Cannot compute noise PSD for unknown basis "{basis}"')
+        return dsets
 
-    # Cut data at start and end
-    if cut_time > 0:
-        n_samples_to_cut = np.round(cut_time * fs).astype(int)
-        new_input_data = input_data[:, :, n_samples_to_cut:-n_samples_to_cut]
-        timestamp = timestamp[n_samples_to_cut:-n_samples_to_cut]
-    else:
-        new_input_data = input_data
+    def run(self, pdata: ProcessedData, inputs: list[str]=None) -> list[str]:
+        # Initialize PSD group in the file if needed
+        if not pdata.has('psd', exact_match=True):
+            psd_group = pdata.create_group('psd')
 
-    # Determine the number of blocks for computing the PSD
-    n_samples = np.size(timestamp)
-    n_samples_per_block = int(2**np.ceil(np.log2(nominal_block_length * fs)))
-    n_blocks = np.floor(float(n_samples) / float(n_samples_per_block)).astype(int)
-    if n_blocks == 0:
-        n_blocks = 1
-        n_samples_per_block = n_samples
-    
-    freq, psd = _compute_psd(new_input_data[:, np.where(chanmask == 1)[0], :], fs, n_samples_per_block)
-    return chanmask, freq, psd
+        psd_group = pdata['psd']
+
+        time = pdata.timestamp[:] - pdata.timestamp[0]
+        bases = self.params['bases']
+        cut_time = self.params['cut_time']
+        nominal_block_length = self.params['nominal_block_length']
+        selection_indices = decode_tone_indices(pdata, self.params['selection_indices'])
+
+        outputs = []
+
+        for basis in bases:
+            match basis:
+                case PsdBasis.IQ:
+                    data = pdata.data_IQ[:]
+                case PsdBasis.GAIN_PHASE:
+                    data = pdata.data_gain_phase[:] / pdata.carrier_amplitude_norm()
+                case PsdBasis.FREQ_DISS:
+                    f = pdata.detector_f()
+                    f[pdata.offres_ind] = 1
+                    data = pdata.data_freq_diss[:] / f[np.newaxis, :, np.newaxis]
+                case _:
+                    raise ValueError(f'{self.name}: Cannot compute noise PSD for unknown basis "{basis}"')
+            if cut_time > 0:
+                n_samples_to_cut = np.round(cut_time * pdata.fs).astype(int)
+                data = data[..., n_samples_to_cut:-n_samples_to_cut]
+                time = time[n_samples_to_cut:-n_samples_to_cut]
+
+            # Determine the number of blocks for computing the PSD
+            n_samples = np.size(time)
+            n_samples_per_block = int(2**np.ceil(np.log2(nominal_block_length * pdata.fs)))
+            n_blocks = np.floor(float(n_samples) / float(n_samples_per_block)).astype(int)
+            if n_blocks == 0:
+                n_blocks = 1
+                n_samples_per_block = n_samples
+
+            # Compute the PSD
+            freq, psd = signal.welch(
+                data[:, selection_indices],
+                pdata.fs,
+                nperseg=n_samples_per_block,
+            )
+
+            if basis in psd_group:
+                del psd_group[basis]
+            basis_group = psd_group.create_group(basis)
+            psd_dset = basis_group.create_dataset('psd', data=psd)
+            outputs.append(psd_dset.name)
+            freq_dset = basis_group.create_dataset('freq', data=freq)
+            outputs.append(freq_dset.name)
+            indices = basis_group.create_dataset('selection_indices', data=selection_indices, dtype=int)
+            outputs.append(indices.name)
+
+        return outputs
 
 
-def _compute_psd(
-        data: npt.NDArray,
-        fs: float,
-        n_samples_per_block: int,
-) -> tuple[npt.NDArray, npt.NDArray]:
-    """Compute the PSD."""
-    return signal.welch(data, fs, nperseg=n_samples_per_block)
+def decode_color_string(color: str) -> tuple[str, str]:
+    match color.lower():
+        case 'b' | 'blue;':
+            med_color = 'b'
+            fill_color = 'cyan'
+        case 'r' | 'red':
+            med_color = 'r'
+            fill_color = 'lightcoral'
+        case 'g' | 'green':
+            med_color = 'g'
+            fill_color = 'lightgreen'
+        case 'k' | 'black':
+            med_color = 'k'
+            fill_color = 'lightgrey'
+        case 'o' | 'orange':
+            med_color = 'darkorange'
+            fill_color = 'bisque'
+        case 'gold':
+            med_color = 'gold'
+            fill_color = 'khaki'
+        case 'turquoise' | 'teal':
+            med_color = 'teal'
+            fill_color = 'turquoise'
+        case 'purple':
+            med_color = 'purple'
+            fill_color = 'violet'
+    return med_color, fill_color
 
 
-@ensure_path(2)
-def plot_psd(
-        freq: npt.NDArray,
-        psd: npt.NDArray,
-        filename: Path,
-        min_percentile: float=16,
-        max_percentile: float=84,
-        title: str | None=None,
-        basis: PsdBasis=PsdBasis.GAIN_PHASE,
-        resonators: list[int]=[0],
-) -> list[Figure]:
-    """Create plots for the psd.
-    
+def plot_psd_df_over_f(
+    freq: npt.NDArray,
+    psd: npt.NDArray,
+    f0: float | None=None,
+    dev_pwr: float | None=None,
+    adc_units_to_hz: float | None=None,
+    csd: npt.NDArray | None=None,
+    offres_median: npt.NDArray | None=None,
+    show_error_band: bool=False,
+    error_band_min_percentile: float=16,
+    error_band_max_percentile: float=84,
+    show_flat_spectrum_level: bool=False,
+    flat_spectrum_search_bounds: tuple[float, float]=(10, 50),
+    title: str | None=None,
+    label: str=None,
+    add_legend: bool=True,
+    ax: plt.Axes=None,
+    figure_kwargs: dict={},
+    freq_color: str='b',
+    diss_color: str='o',
+    offres_color: str='r',
+    xlim: tuple[float, float]=None,
+    ylim: tuple[float, float]=None,
+    title_fontsize: int=16,
+    label_fontsize: int=16,
+    tick_size: int=14,
+) -> Figure | None:
+    """Plot df/f noise for a single resonator.
+
     Args:
         freq (npt.NDArray): Array of frequencies (N_freq).
-        psd: (npt.NDArray): PSD (N_chan x N_resonators x N_freq).
-        filename (Path): PDF filename to save the  plots to.
-        min_perncentile (float, optional): Percentile of lower error bound for the plot.
+        psd: (npt.NDArray): Frequency / disspiation PSD (2 x N_freq) or (2 x N_tones x N_freq).
+        show_error_band (bool, optional): Whether to show the error band. Defaults
+            to True.
+        error_band_min_perncentile (float, optional): Percentile of lower error bound for the plot.
             Defaults to 16.
-        max_perncentile (float, optional): Percentile of upper error bound for the plot
+        error_band_max_perncentile (float, optional): Percentile of upper error bound for the plot
             Defaults to 84.
-        title (str, optional): Title to give to each plot. Defaults to None.
-        basis (str, optional): The basis of the data. Either IQ ('iq'), 
-            Gain/Phase ('gp'), or Frequency/Dissipation ('fd'). Defaults to 'gp.'
-    
+        title (str, optional): Title to give to the plot. Defaults to None.
+        `ax` (plt.Axes, optional): Axes to plot in.
+
     Returns:
-        (list[Figure]): N_chan + 1 plots corresponding to the PSD along the
-            first basis direction, second direction, and mean across both directions.
-    
+        Figure | None: If no `ax` was provided, a new figure is generated to
+            create the plot and is returned.
+
     Raises:
         ValueError: If `basis` is not a valid basis (see `VALID_BASES`).
     """
+    fig = None
 
-    # cutoff = 250  # Number of data points to cut off at the end
-    # psd = psd[:, :, :-cutoff]
-    # freq = freq[:-cutoff]
+    # Create figure if needed
+    if ax is None:
+        fig = plt.figure(**figure_kwargs)
+        ax = fig.add_subplot()
 
-    if title is None:
-        title = 'RFSoC Loopback PSD'
-    match basis:
-        case PsdBasis.GAIN_PHASE:
-            titles = [[f'Tile {i_chan} {title} - Gain', f'Tile {i_chan} {title} - Phase'] for i_chan in range(psd.shape[0])]
-            ylabel = r'Noise PSD ($\text{dBc Hz}^{-1})$'
-            yscale = 'linear'
-        case PsdBasis.IQ:
-            titles = [[f'Tile {i_chan} {title} - I', f'Tile {i_chan} {title} - Q'] for i_chan in range(psd.shape[0])]
-            ylabel = r'Noise PSD ($\text{dBc Hz}^{-1})$'
-            yscale = 'linear'
-        case PsdBasis.FREQ_DISS:
-            return plot_psd_df_over_f(freq, psd, filename, title=title, resonators=resonators)
-        case _:
-            raise ValueError(f'Invalid basis {basis}; must be one of {VALID_BASES}')
+        # Setup plot
+        ax.set_xscale('log')
+        ax.set_yscale('log')
+        if xlim is not None:
+            ax.set_xlim(*xlim)
+        if ylim is not None:
+            ax.set_ylim(*ylim)
+        ax.set_xlabel('Frequency (Hz)', fontsize=label_fontsize)
+        ax.set_ylabel(r'Sdf/f ($Hz^{-1}$)', fontsize=label_fontsize)
+        ax.tick_params(labelsize=tick_size)
+        if f0 is not None:
+            title += f' (f0 = {f0 / 1e6:.3f} MHz)'
+        if dev_pwr is not None:
+            title += f' at dev_pwr {dev_pwr}'
+        if title is not None:
+            ax.set_title(title, fontsize=title_fontsize)
 
-    n_plots = psd.shape[1]
-    psd_med = np.median(psd, axis=2)
+    # Select color
+    med_color_freq, fill_color_freq = decode_color_string(freq_color)
+    med_color_diss, fill_color_diss = decode_color_string(diss_color)
+    med_colors = [med_color_freq, med_color_diss]
+    fill_colors=  [fill_color_freq, fill_color_diss]
 
-    plot_data_med = 10 * np.log10(psd_med)
+    super_labels = ['Frequency', 'Dissipation']
+    labels = [' - '.join(filter(None, (super_label, label))) for super_label in super_labels]
 
-    psd_min = psd_med[:]
-    psd_max = psd_med[:]
+    if psd.ndim == 3:
+        psd_med = np.median(psd, axis=1)
+        plot_data_med = psd_med
+    else:
+        plot_data_med = psd
 
-    if psd.shape[1] > 1:
-        psd_min = np.percentile(psd, min_percentile, axis=2)
-        psd_max = np.percentile(psd, max_percentile, axis=2)
-
-    plot_data_min = 10 * np.log10(psd_min)
-    plot_data_max = 10 * np.log10(psd_max)
-
-    # Plot 
-    if not filename.exists():
-        filename.touch(PERMISSIONS_ALL_FULL)
-    figs = []
-    with PdfPages(filename) as pdf:
-        for i_chan in range(psd.shape[0]):
-            for i_plot in range(n_plots):
-                fig = create_plot(
-                    freq,
-                    plot_data_min[i_chan, i_plot],
-                    plot_data_med[i_chan, i_plot],
-                    plot_data_max[i_chan, i_plot],
-                    percentiles=(min_percentile, max_percentile),
-                    title=titles[i_chan][i_plot],
-                    ylabel=ylabel,
-                    yscale=yscale,
-                )
-                pdf.savefig(fig)
-                figs.append(fig)
-        average_fig = create_plot(
-            freq,
-            np.sum(plot_data_min[i_chan], axis=0) / n_plots,
-            np.sum(plot_data_med[i_chan], axis=0) / n_plots,
-            np.sum(plot_data_max[i_chan], axis=0) / n_plots,
-            percentiles=(min_percentile, max_percentile),
-            title= f'Tile {i_chan} {title} - Averaged',
-            ylabel=ylabel,
-            yscale=yscale,
+    # Flat spectrum level
+    if show_flat_spectrum_level:
+        flat_spectrum_idx = np.where(
+            (freq > flat_spectrum_search_bounds[0]) &
+            (freq < flat_spectrum_search_bounds[1])
         )
-        pdf.savefig(average_fig)
+        new_labels = []
+        flat_spectrum_noise_levels = np.zeros(2)
+        for j, this_label in enumerate(labels):
+            flat_spectrum_noise_levels[j] = np.median(plot_data_med[j, ..., flat_spectrum_idx])
+            new_labels.append(rf'{this_label} ({flat_spectrum_noise_levels[j]:.1e} Hz$^{{-1}}$)')
+        labels = new_labels
 
-    return figs
+    # Plot PSD
+    for j, this_label in enumerate(labels):
+        # Error band
+        if show_error_band:
+            if psd.ndim != 3:
+                _logger.error('Cannot show error band for PSD with dimensions != 3.')
+            else:
+                psd_min = np.percentile(psd[j], error_band_min_percentile, axis=0)
+                psd_max = np.percentile(psd[j], error_band_max_percentile, axis=0)
+                ax.fill_between(
+                    freq,
+                    psd_min,
+                    psd_max,
+                    facecolor=fill_colors[j],
+                    alpha=0.5,
+                )
 
-@ensure_path(2)
-def plot_psd_df_over_f(
-        freq: npt.NDArray,
-        psd: npt.NDArray,
-        filename: Path,
-        title: str | None=None,
-        resonators: list[int]=[0],
-) -> list[Figure]:
+        # Flat spectrum level
+        if show_flat_spectrum_level:
+            ax.axhline(
+                flat_spectrum_noise_levels[j],
+                color=med_colors[j],
+                linestyle='dashed',
+            )
+        ax.plot(freq, plot_data_med[j], color=med_colors[j], label=this_label)
+        if offres_median is not None:
+            ax.plot(freq, offres_median[j], linestyle='dashed', color=offres_color, label=f'Off-Resonance {super_labels[j]} Median')
+
+    # Add legend
+    if add_legend:
+        ax.legend(fontsize=14)
+
+    if fig is not None:
+        fig.tight_layout()
+        return fig
+
+
+def plot_psd_dbc_hz(
+    freq: npt.NDArray,
+    psd: npt.NDArray,
+    xlim: tuple[float, float]=None,
+    ylim: tuple[float, float]=None,
+    show_error_band: bool=True,
+    error_band_min_percentile: float=16,
+    error_band_max_percentile: float=84,
+    show_flat_spectrum_level: bool=True,
+    flat_spectrum_search_bounds: tuple[float, float]=(10, 50),
+    label: str=None,
+    title: str | None=None,
+    add_legend: bool=True,
+    ax: plt.Axes=None,
+    figure_kwargs: dict={},
+    color: str='b',
+) -> Figure | None:
     """Create plots for the psd.
-    
+
     Args:
         freq (npt.NDArray): Array of frequencies (N_freq).
-        psd: (npt.NDArray): PSD (N_chan x N_resonators x N_freq).
-        filename (Path): PDF filename to save the  plots to.
-        title (str, optional): Title to give to each plot. Defaults to None.
-    
+        psd: (npt.NDArray): PSD (N_tones x N_freq).
+        show_error_band (bool, optional): Whether to show the error band. Defaults
+            to True.
+        error_band_min_perncentile (float, optional): Percentile of lower error bound for the plot.
+            Defaults to 16.
+        error_band_max_perncentile (float, optional): Percentile of upper error bound for the plot
+            Defaults to 84.
+        title (str, optional): Title to give to the plot. Defaults to None.
+        `ax` (plt.Axes, optional): Axes to plot in.
+
     Returns:
-        (list[Figure]): N_chan plots corresponding to the PSD for each resonator.
-    
+        Figure | None: If no `ax` was provided, a new figure is generated to
+            create the plot and is returned.
+
     Raises:
-        ValueError: If the length of `resonators` is greater than the number of resonators
+        ValueError: If `basis` is not a valid basis (see `VALID_BASES`).
     """
-    ylabel = r'Sdf/f ($Hz^{-1}$)'
-    # Plot 
-    if not filename.exists():
-        filename.touch(PERMISSIONS_ALL_FULL)
-    figs = []
-    with PdfPages(filename) as pdf:
-        for i, res in enumerate(resonators):
-            res_title = title + f' - Resonator {res}'
-            fig = plot_df_over_f(
-                freq,
-                psd[:, i, :],
-                ylabel=ylabel,
-                title=res_title,
-            )
-            pdf.savefig(fig)
-            plt.close(fig)
-            figs.append(fig)
-        return figs
+    fig = None
 
+    # Create figure if needed
+    if ax is None:
+        fig = plt.figure(**figure_kwargs)
+        ax = fig.add_subplot()
 
-def plot_df_over_f(
-    x_data: npt.NDArray,
-    y_data: npt.ArrayLike,
-    title: str | None=None,
-    ylabel: str='Noise PSD (df / f)',
-) -> Figure:
-    """Create a plot of the noise PSD in df/f units."""
-    fig = plt.figure(figsize=(9, 6))
-    ax = plt.subplot()
-    for j, label in enumerate(['Frequency', 'Dissipation']):
-        ax.plot(x_data, y_data[j], label=label)
-    ax.set_xscale('log')
-    ax.set_xlim(1, 250)
-    ax.set_yscale('log')
-    # ax.set_ylim(1e-17,1e-15)
+        # Setup plot
+        ax.set_xscale('log')
+        ax.set_yscale('linear')
+        if xlim is not None:
+            ax.set_xlim(*xlim)
+        if ylim is not None:
+            ax.set_ylim(*ylim)
+        ax.set_xlabel('Frequency (Hz)', fontsize=16)
+        ax.set_ylabel(r'Noise PSD (dBc/Hz)', fontsize=16)
+        ax.tick_params(labelsize=14)
+        if title is not None:
+            ax.set_title(title, fontsize=16)
 
-    ax.set_xlabel('Frequency (Hz)', fontsize=16)
-        
-    ax.set_ylabel(ylabel, fontsize=16)
-    ax.tick_params(labelsize=14)
-    ax.legend(fontsize=14, loc='best')
-    if title is not None:
-        ax.set_title(title, fontsize=16)
-    plt.tight_layout()
-    return fig
+    # Select color
+    match color.lower():
+        case 'b' | 'blue;':
+            med_color = 'b'
+            fill_color = 'cyan'
+        case 'r' | 'red':
+            med_color = 'r'
+            fill_color = 'lightcoral'
+        case 'g' | 'green':
+            med_color = 'g'
+            fill_color = 'lightgreen'
+        case 'k' | 'black':
+            med_color = 'k'
+            fill_color = 'lightgrey'
+        case 'o' | 'orange':
+            med_color = 'darkorange'
+            fill_color = 'bisque'
+        case 'gold':
+            med_color = 'gold'
+            fill_color = 'khaki'
+        case 'turquoise' | 'teal':
+            med_color = 'teal'
+            fill_color = 'turquoise'
+        case 'purple':
+            med_color = 'purple'
+            fill_color = 'violet'
 
+    # Plot median
+    psd_med = np.median(psd, axis=0)
+    plot_data_med = 10 * np.log10(psd_med)
 
-def create_plot(
-    xdata: npt.ArrayLike,
-    ydata_min: npt.ArrayLike,
-    ydata_med: npt.ArrayLike,
-    ydata_max: npt.ArrayLike,
-    percentiles: tuple[float, float]=(16., 84.),
-    label: str='Median Measured Noise',
-    title: str | None=None,
-    xlabel: str='Frequency (Hz)',
-    ylabel: str=r'Noise PSD (dBc/Hz)',
-    yscale: str='linear',
-) -> Figure:
-    """Create a plot of the noise PSD."""
-    fig = plt.figure(figsize=(9, 6))
-    ax = plt.subplot()
-    ax.plot(xdata, ydata_med, color='b', label=label)
-    flat_spectrum_idx = np.where((xdata > 10) & (xdata < 50))
-    flat_spectrum_noise = np.median(ydata_med[flat_spectrum_idx])
-    plt.hlines(flat_spectrum_noise, XLIM[0], XLIM[1], colors='r', linestyles='dashed', label=f'Flat Spectrum Level = {flat_spectrum_noise:.1f} dBc/Hz')
-    ax.fill_between(
-        xdata,
-        ydata_min,
-        ydata_max,
-        facecolor='c',
-        alpha=0.5,
-        label=f'{ordinal(int(percentiles[0]))} Percentile to {ordinal(int(percentiles[1]))} Percentile'
-    )
-    ax.set_xscale('log')
-    ax.set_xlim(*XLIM)
-    ax.set_yscale(yscale)
-    if yscale=='linear' and np.median(ydata_min) > -110 and np.median(ydata_max) < -60:
-        ax.set_ylim(-110, -60)
-        loc = 'upper right'
+    # Error band
+    if show_error_band:
+        psd_min = np.percentile(psd, error_band_min_percentile, axis=0)
+        psd_max = np.percentile(psd, error_band_max_percentile, axis=0)
+        plot_data_min = 10 * np.log10(psd_min)
+        plot_data_max = 10 * np.log10(psd_max)
+        ax.fill_between(
+            freq,
+            plot_data_min,
+            plot_data_max,
+            facecolor=fill_color,
+            alpha=0.5,
+        )
+
+    # Flat spectrum level
+    if show_flat_spectrum_level:
+        flat_spectrum_idx = np.where(
+            (freq > flat_spectrum_search_bounds[0]) &
+            (freq < flat_spectrum_search_bounds[1])
+        )
+        flat_spectrum_noise_level = np.median(plot_data_med[flat_spectrum_idx])
+        lines = ax.plot(
+            freq,
+            plot_data_med,
+            color=med_color,
+        )
+        if label is not None:
+            lines[0].set_label(rf'{label} ({flat_spectrum_noise_level:.1f} dBc Hz$^{{-1}}$)')
+        ax.axhline(
+            flat_spectrum_noise_level,
+            color=med_color,
+            linestyle='dashed',
+        )
     else:
-        loc = 'best'
-    
-    ax.set_xlabel(xlabel, fontsize=16)
-    ax.set_ylabel(ylabel, fontsize=16)
-    ax.tick_params(labelsize=14)
-    ax.legend(fontsize=14, loc=loc)
-    if title is None:
-        title = 'RFSoC Loopback PSD'
-    ax.set_title(title, fontsize=16)
-    plt.tight_layout()
+        lines = ax.plot(freq, plot_data_med, color='b', label=label)
 
-    return fig
+    # Add legend
+    if add_legend and label is not None:
+        ax.legend(fontsize=14)
+
+    if fig is not None:
+        fig.tight_layout()
+        return fig
 
 
-# if __name__ == '__main__':
-#     pairs = [
-#         # ('equal_0-256', 'RFSoC Loopback with 1000 Tones Over Full Bandwidth'),
-#         # ('equal_1-255', 'RFSoC Loopback with 1000 Tones in Range +/-[1, 255] MHz'),
-#         # ('equal_5-251', 'RFSoC Loopback with 1000 Tones in Range +/-[5, 251] MHz'),
-#         # ('equal_10-246', 'RFSoC Loopback with 1000 Tones in Range +/-[10, 246] MHz'),
-#         # ('data/default_0-256.hdf5', 'RFSoC Loopback with Default Tones'),
-#         # ('default_1-255', 'RFSoC Loopback with Default Tones in Range +/-[1, 255] MHz'),
-#         # ('default_5-251', 'RFSoC Loopback with Default Tones in Range +/-[5, 251] MHz'),
-#         # ('./data/default_10-246', 'RFSoC Loopback with Default Tones in Range +/-[10, 246] MHz'),
-#         # ('/data/20250404/20250404_chan_1_TOD_set1001.h5', 'ASU Readout'),
-#         ('/data/20250415/20250415_chan_1_TOD_set1004.h5', 'Loopback')
-#     ]
-#     for name, title in pairs:
-#         # input_data, timestamp, chanmask = load_time_ordered_IQ_data(f'data/{name}.hdf5')
-#         input_data, timestamp, chanmask = load_time_ordered_IQ_data(f'{name}')
-#         # input_data = input_data[:, :-5, :]
-#         rotated_data = rotate_to_amplitude_and_phase(input_data)
-#         # save_name = f'new_psd_{name}'
-#         save_name = Path(name).name
+@register_routine
+class PlotPSD(DataRoutine):
+    name = 'PlotPSD'
+    version = '1.0.0'
 
-#         chanmask, freq, noise_psd = compute_noise_psd(
-#             rotated_data,
-#             timestamp,
-#             chanmask=None,
-#             ds_factor=3,
-#             flag_outliers=True,
-#             nominal_block_length=10,
-#             outlier_sigma=2,
-#         )
-#         plot_psd(freq, noise_psd, f'plots/{save_name}.pdf', basis='gp', title=title)
-#         plt.close()
+    def __init__(
+            self,
+            *bases: PsdBasis,
+            min_percentile: float=16,
+            max_percentile: float=84,
+            title: str=None,
+            show: bool=False,
+    ):
+        super().__init__(
+            bases=bases,
+            min_percentile=min_percentile,
+            max_percentile=max_percentile,
+            title=title,
+            show=show,
+        )
 
+    def inputs(self, pdata: ProcessedData) -> list[str]:
+        dsets = []
+        bases = self.params['bases']
+        for basis in bases:
+            if basis not in PsdBasis:
+                raise ValueError(f'{self.name}: Unknown PSD basis "{basis}"')
+            dsets.append(f'/psd/{basis}/psd')
+            dsets.append(f'/psd/{basis}/freq')
+            if basis == PsdBasis.FREQ_DISS:
+                dsets.append(f'/psd/{basis}/selection_indices')
+        return dsets
 
-if __name__ == '__main__':
-    parser = ArgumentParser(description='Compute the noise PSD from RFSoC data.')
-    parser.add_argument('date', type=str, help='Date of the data in YYYYMMDD format.')
-    parser.add_argument('setnum', type=int, help='Set number of the data.')
-    parser.add_argument('--outlier_sigma', type=float, default=2.0, help='Sigma for outlier detection.')
-    parser.add_argument('-d', '--ds_factor', type=int, default=1, help='Downsampling factor.')
-    parser.add_argument('-f', '--do_flag_outliers', action='store_true', help='Flag outliers in the data.')
-    parser.add_argument('-n', '--remove_noise', action='store_true', help='Remove electronics noise from the data.')
-    parser.add_argument('--lp_filt_freq', type=float, default=10, help='Low-pass filter frequency in HZ for electronics noise removal (defaults to 10).')
-    parser.add_argument('-b', '--basis', type=str, choices=VALID_BASES, default='gp', help='Basis of the data (gp, iq, fd).')
-    parser.add_argument('-p', '--show_plots', action='store_true', help='Show noise plots to screen when finished.')
-    parser.add_argument('--block_length', type=float, default=10, help='Nominal block length. Time in seconds for a single "block" of data (defaults to 10s).')
-    parser.add_argument('--cut_time', type=float, default=10, help='Time in seconds to cut from teh ends of the data (defaults to 10).')
-    parser.add_argument('--title', type=str, default='Noise PSD', help='Title to use for the plots')
-    parser.add_argument('-o', '--output', type=str, default='', help='Output filename (defaults to DATE_setSETNUM_psd_BASIS_TITLE.pdf).')
-    parser.add_argument('--max_eigenmodes', type=int, default=30, help='Maximum number of eigenmodes to use for electronics noise removal (defaults to 30).')
-    args = parser.parse_args()
+    def run(self, pdata: ProcessedData, inputs: list[str]=None) -> list[str]:
+        bases = self.params['bases']
+        title = self.params['title']
+        if title is None:
+            title = ''
+        for basis in bases:
+            pdf_path = f'{pdata.file_stub}_psd_{basis}.pdf'
+            basis_group = pdata[f'psd/{basis}']
+            match basis:
+                case PsdBasis.IQ | PsdBasis.GAIN_PHASE:
+                    if basis == PsdBasis.IQ:
+                        subtitles = ['I', 'Q', 'Average']
+                    else:
+                        subtitles = ['Gain', 'Phase', 'Average']
+                    titles = list(' - '.join(filter(None, (title, subtitle))) for subtitle in subtitles)
+                    with PdfPages(pdf_path) as pdf:
+                        fig0 = plot_psd_dbc_hz(
+                            basis_group['freq'][:],
+                            basis_group['psd'][0],
+                            title=titles[0],
+                        )
+                        fig1 = plot_psd_dbc_hz(
+                            basis_group['freq'][:],
+                            basis_group['psd'][1],
+                            title=titles[1],
+                        )
+                        fig2 = plot_psd_dbc_hz(
+                            basis_group['freq'][:],
+                            np.mean(basis_group['psd'], axis=0),
+                            title=titles[2],
+                        )
+                        pdf.savefig(fig0)
+                        pdf.savefig(fig1)
+                        pdf.savefig(fig2)
+                        if self.params['show']:
+                            plt.show()
+                        plt.close(fig0)
+                        plt.close(fig1)
+                        plt.close(fig2)
+                case PsdBasis.FREQ_DISS:
+                    tones = basis_group['selection_indices'][:]
+                    freq = basis_group['freq'][:]
+                    psd = basis_group['psd'][:]
+                    detector_f = pdata.detector_f()
 
-    date = args.date
-    setnum = args.setnum
-    outlier_sigma = args.outlier_sigma
-    ds_factor = args.ds_factor
-    do_flag_outliers = args.do_flag_outliers
-    remove_noise = args.remove_noise
-    basis = args.basis
-    nominal_block_length = args.block_length
-    cut_time = args.cut_time
-    lp_filt_freq = args.lp_filt_freq
-    max_modes = args.max_eigenmodes
-    title = args.title
-    if args.output == '':
-        output_file = f'{DEFAULT_DATA_DIRECTORY}/{date}/{date}_set{setnum}_psd_{basis}_{title}.pdf'
-    else:
-        output = args.output
+                    onres_tones = np.intersect1d(tones, pdata.onres_ind)
+                    onres_indices = np.isin(tones, onres_tones).nonzero()
+                    onres_psd = psd[:, onres_indices[0]]
+                    offres_tones = np.intersect1d(tones, pdata.offres_ind)
+                    if offres_tones.size != 0:
+                        offres_indices = np.isin(tones, offres_tones).nonzero()
+                        offres_median = np.median(psd[:, offres_indices[0]], axis=1)
+                    else:
+                        offres_median = None
+                    with PdfPages(pdf_path) as pdf:
+                        # Plot all on-resonance tones
+                        onres_fig = plot_psd_df_over_f(
+                            freq,
+                            onres_psd,
+                            title=' - '.join(filter(None, (title, f'On-Resonance Tones'))),
+                            show_error_band=True,
+                            add_legend=True,
+                            show_flat_spectrum_level=True,
+                        )
+                        pdf.savefig(onres_fig)
+                        plt.close(onres_fig)
 
-    pd = ProcessedData.from_tod(
-        date,
-        setnum,
-        do_electronics_noise_removal=remove_noise,
-        max_modes=max_modes,
-        ds_factor=ds_factor,
-        electronics_noise_lp_filt_freq=lp_filt_freq,
-    )
+                        # Plot individual tones
+                        for i, i_tone in enumerate(tones):
+                            f0 = detector_f[i_tone]
+                            if offres_median is not None:
+                                this_offres_median = offres_median / (pdata.adc_units_to_hz[i_tone] * f0) ** 2
+                            else:
+                                this_offres_median = None
+                            fig = plot_psd_df_over_f(
+                                freq,
+                                psd[:, i],
+                                f0=detector_f[i_tone],
+                                offres_median=this_offres_median,
+                                title=' - '.join(filter(None, (title, f'Resonator {i_tone}'))),
+                                add_legend=True,
+                                show_flat_spectrum_level=True,
+                            )
+                            pdf.savefig(fig)
+                            plt.close(fig)
+                        
+        return []
 
-
-
-    match basis:
-        case 'iq':
-            # IQ basis
-            # input_data = rotate_basis(p.data_gain_phase, -p.IQ_to_gain_phase_angle)
-            input_data = pd.data_IQ
-        case 'fd':
-            # Frequency/Dissipation basis
-            # Get frequencies from the raw data file
-            # raw_data_file = f'/data/{date}/{date}_chan_1_TOD_set{setnum}.h5'
-            # TODO: Get the channel name from the processed data file? (but there's multiple in theory??)
-            raw_data_file = get_tod_template(date, setnum, chan_name='Be231102p2_100_tones')
-            # raw_data_file = get_tod_template(date, setnum, chan_name='1000_tone_uniform_202050829')
-            fh = RawDataFile(raw_data_file, 'r')
-            freq = fh.baseband_freqs[:] + fh.lo_freq[:]
-
-            input_data = pd.data_freq_diss / freq[np.newaxis, :, np.newaxis]
-        case 'gp':
-            # Gain/Phase basis
-            input_data = pd.data_gain_phase / pd.carrier_amplitude_norm()
-        case _:
-            raise ValueError(f'Invalid basis {basis}; must be one of {VALID_BASES}')
-
-    # Fix data dimensions
-    first_dimension = input_data.shape[0] if input_data.ndim == 3 else 1
-    if first_dimension == 1:
-        input_data = input_data.reshape((1, *input_data.shape))
-
-    chanmask = pd.chanmask[:]
-
-    # Flag outliers
-    if do_flag_outliers:
-        chanmask = flag_outliers(input_data, pd.fs, chanmask, sigma=outlier_sigma)
-
-    filt_sos = signal.butter(6, 1, btype='highpass', fs=pd.fs, output='sos', analog=False)
-    input_data[:] = signal.sosfiltfilt(filt_sos, input_data)
-
-    chanmask, freq, noise_psd = compute_noise_psd(
-        input_data,
-        pd.timestamp,
-        chanmask=chanmask,
-        nominal_block_length=nominal_block_length,
-        cut_time=cut_time,
-    )
-    plot_psd(
-        freq,
-        noise_psd,
-        output_file,
-        basis=basis,
-        title=title,
-        resonators=np.where(chanmask==1)[0],
-    )
-    if args.show_plots:
-        plt.show()
-    pd.close()
