+++ rfsocinterface/core/data/__init__.py
@@ -12,19 +12,15 @@ from rfsocinterface.core.data.pipeline import Pipeline
 from rfsocinterface.core.data.routines import (
     ROUTINE_REGISTRY,
     CleanTOD,
-    ComputeNoisePSD,
     CutoffFilter,
     DataRoutine,
     HighPassFilter,
     LowPassFilter,
-    PlotPSD,
     ProcessingStage,
     RemoveElectronicsNoise,
     compute_templates,
     decode_tone_indices,
     find_peaks,
-    plot_psd_dbc_hz,
-    plot_psd_df_over_f,
     register_routine,
 )
 from rfsocinterface.core.data.storage import (
@@ -55,47 +51,43 @@ from rfsocinterface.core.data.utils import (
 ROUTINE_NAME_MAP = ROUTINE_REGISTRY  # Temporary fix until further edits made
 
 __all__ = [
-    AZ_TRIM,
-    ZA_TRIM,
-    CALIBRATION_TABLE_DTYPE,
-    DEFAULT_MAP_DPIX,
-    DECIMATE_ORDER,
-    N_POLARIZATION,
-    OPTCAM_OFFSET_AZ_PIX,
-    OPTCAM_OFFSET_ZA_PIX,
-    OPTCAM_PIX_SIZE_DEGREES,
-    ROUTINE_REGISTRY,
-    RFSOC_TIME_OFFSET,
-    TONES_TABLE_DTYPE,
-    BinTODIntoMap,
-    CleanTOD,
-    ComputeNoisePSD,
-    ConsolidatedData,
-    CutoffFilter,
-    DataRoutine,
-    HighPassFilter,
-    LowPassFilter,
-    MakeVideo,
-    NewDataStorage,
-    Pipeline,
-    PlotMap,
-    PlotPSD,
-    ProcessedData,
-    ProcessingStage,
-    PsdBasis,
-    RemoveElectronicsNoise,
-    compute_templates,
-    decode_tone_indices,
-    find_peaks,
-    flag_outliers,
-    generate_calibrated_data,
-    get_channel_group_name,
-    get_extent,
-    get_scaled_optical_image,
-    get_step_group_name,
-    plot_map,
-    plot_psd_dbc_hz,
-    plot_psd_df_over_f,
-    register_routine,
-    rotate_basis,
+    'AZ_TRIM',
+    'ZA_TRIM',
+    'CALIBRATION_TABLE_DTYPE',
+    'DEFAULT_MAP_DPIX',
+    'DECIMATE_ORDER',
+    'N_POLARIZATION',
+    'OPTCAM_OFFSET_AZ_PIX',
+    'OPTCAM_OFFSET_ZA_PIX',
+    'OPTCAM_PIX_SIZE_DEGREES',
+    'ROUTINE_REGISTRY',
+    'RFSOC_TIME_OFFSET',
+    'TONES_TABLE_DTYPE',
+    'BinTODIntoMap',
+    'CleanTOD',
+    'ConsolidatedData',
+    'CutoffFilter',
+    'DataRoutine',
+    'HighPassFilter',
+    'LowPassFilter',
+    'MakeVideo',
+    'NewDataStorage',
+    'Pipeline',
+    'PlotMap',
+    'ProcessedData',
+    'ProcessingStage',
+    'PsdBasis',
+    'RemoveElectronicsNoise',
+    'compute_templates',
+    'decode_tone_indices',
+    'find_peaks',
+    'flag_outliers',
+    'generate_calibrated_data',
+    'get_channel_group_name',
+    'get_extent',
+    'get_scaled_optical_image',
+    'get_step_group_name',
+    'plot_map',
+    'register_routine',
+    'rotate_basis',
 ]
